+++ ThinkingHome.Plugins.Tmp/Resources/tmp.js
@@ -1,8 +1,10 @@
 var lib = require('lib');
-var template = '<h1>Hi!</h1><p>This is tmp page.</p>';
+var template = '<h1>Hi!</h1><p>This is tmp page.</p><p>{{lang \'web-test\'}}</p><p>{{lang \'web-test-2\'}}</p><p>{{lang \'web-test-3\'}}</p>';
+var lang = require('lang!static/tmp/lang.json');
 
 var View = lib.marionette.View.extend({
-    template: lib.handlebars.compile(template)
+    template: lib.handlebars.compile(template),
+    templateContext: { lang: lang }
 });
 
 var Section = lib.common.AppSection.extend({
@@ -10,7 +12,7 @@ var Section = lib.common.AppSection.extend({
         var view = new View();
         this.application.setContentView(view);
 
-        return lib.ajax.getJSON('/qwdqwkfqwgqwgqwgqwgqwg');
+        //return lib.ajax.getJSON('/qwdqwkfqwgqwgqwgqwgqwg');
     }
 });