+++ ThinkingHome.Plugins.WebUi/Resources/Application/lib.js
@@ -28,6 +28,14 @@ define([
         start: function () { }
     });
 
+    var stringLocalizer = function(data) {
+        this.data = data;
+    };
+
+    stringLocalizer.prototype.get = function(key, defaultValue) {
+        return (this.data && this.data[key]) || defaultValue || key;
+    };
+
     //region ajax
 
 
@@ -124,7 +132,7 @@ define([
         value = value + '';
         length = length || 0;
 
-        return (Array(length + 1).join('0') + value).slice(-length);
+        return (new Array(length + 1).join('0') + value).slice(-length);
     });
 
     handlebars.default.registerHelper('isnull', function(variable, options) {
@@ -135,12 +143,17 @@ define([
         }
     });
 
+    handlebars.default.registerHelper('lang', function (key) {
+        return this.lang instanceof stringLocalizer ? this.lang.get(key) : key;
+    });
+
     //endregion
 
     return {
         common: {
             ApplicationBlock: applicationBlock,
-            AppSection: appSection
+            AppSection: appSection,
+            StringLocalizer: stringLocalizer
         },
         form: {
             setOptions: setOptions