+++ ThinkingHome.Plugins.WebUi/Resources/Application/loaders/system-tpl.js
@@ -1,9 +1,0 @@
-var lib = require('lib');
-
-exports.translate = function (load) {
-    var precompiled = lib.handlebars.precompile(load.source);
-
-    load.metadata.format = "amd";
-
-    return 'define([\'lib\'], function (lib) {\n    return lib.handlebars.template(' + precompiled + ');\n});';
-};